+++ Changes.md
@@ -9,7 +9,7 @@
 Features
 --------
 
-- TaskNode : Added `isolate` plug. When an isolated task is dispatched, a script containing only the task node will be saved for each dispatched batch and all plug inputs will be computed at dispatch time.
+- TaskNode : Added `isolate` plug. When an isolated task is dispatched, a script containing only the task node will be saved for each dispatched batch and all plug inputs will be computed at dispatch time. Plugs with values that vary based on the task batch frames are connected to a spreadsheet containing a row per-frame and a column per-plug.
 
 Improvements
 ------------